requirements.txt
Flask==3.0.3
flask-cors==5.0.0
numpy==1.26.4
opencv-python==4.10.0.84
colorthief==0.2.1
Pillow==9.3.0
